Update to Seeker
Includes
- Index Searching
- Keybind Search Paths
- Folder Depth integer
- Updated UI

Note: This was designed for my purpose and may need to be changed, such as the Default Search Path and Favicon.
The intended use was to work from a server for 100's of people to use the same "app" without the need to synchronize updates. Since everyone runs the app, they are always updated once the script runs. Zero Installation.
Will need to create a shortcut with the Target command.
`C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe -ExecutionPolicy Bypass -WindowStyle Hidden -File "[Replace with Path]\Seeker.ps1"`